Finished train.py and evaluated the first models. Next should be to do some feature engineering, use a different model than FNN and do some visualisation.

+++ diabetes_prediction/dataset.py
@@ -16,8 +16,8 @@ class DiabetesDataset(Dataset):
             data : np.array,
             targets : np.array
     ):
-        self.data = torch.tensor(data.astype(torch.float32))
-        self.targets = torch.tensor(targets.astype(torch.long))
+        self.data = torch.tensor(data, dtype=torch.float32)
+        self.targets = torch.tensor(targets, dtype=torch.float32)
 
     def __len__(self):
         return len(self.data)
